Update project name and database configurations
The project name's been updated from Projecto to Atlas in all configurations in the Dockerfile, and the database server name in the application's production settings (appsettings.Production.json) from db to atlas-db. The changes in docker-compose.yml mostly reflects the renaming of the app service and db service. This commit was necessary due to the renaming of the project and its associated databases.