Replace vectormap-data for Eurasia with Iran
The vectormap-data for Eurasia in the Shared Layout was replaced with that for Iran. This change was necessary to update the geographical information displayed on the vector map. Moreover, a corresponding JavaScript file for Iran was added to ensure the map data can properly be loaded and rendered in the application.